training/requirements.txt
kfp==1.8.9
google-cloud-pipeline-components==0.2.0
dask[dataframe]==2021.12.0
tensorflow-data-validation
tensorflow
apache-beam==2.35.0
pytest==6.2.5
pytest-mock==3.6.1
pytest-cov==3.0.0
google-cloud-bigquery==2.24.1
tensorflow-data-validation==1.5.0
apache-beam==2.35.0
gcsfs==2021.11.1
scikit-learn==0.24.0
pandas==1.3.5
numpy==1.19.5
fsspec==2021.11.1
gcsfs==2021.11.1